fix: convert start_indices to int in Lorenz63Dataset
window_spacing can be float (e.g. evaluate_all.py passes 3.0),
causing np.arange * float to produce float64 indices that break
tensor slicing (idx + cfg.num_steps becomes float).

+++ data/lorenz63.py
@@ -154,7 +154,7 @@ class Lorenz63Dataset:
 
         start_indices = (
             np.arange(cfg.num_windows) * cfg.window_spacing + cfg.spinup_steps
-        )
+        ).astype(int)
 
         self.windows = []
         for idx in start_indices:
